Add SMMUv2 support
This driver is used to setup SMMU mappings for IOspaces.  For simplicity
it uses one Stream Mapping Register per Context Bank, but this limits
the number of available IOspaces to the lesser of the number of available
stream mapping registers or context banks.

For devices with more DMA capable devices than this limit it may
not be possible to statically allocate enough IOspaces to cover all
potential DMA masters, so stream mapping/context bank pairs are
initialized to sane defaults and allocated as IOspaces are created.
If an application needs more DMA capable devices than this, the driver
and supporting code will need to be updated to more intelligently
group DMA masters by some criteria, such as VMID or a mask on their
stream IDs.

Change-Id: Ie85578459afe778829458be49a1b48bf0c41ce4c