Updated vspace to reflect MMU changes when hypervisor support is enabled.
The MMU functions differently when Hypervisor support is enabled.
Virtualization extensions become enabled, and cause the userspace tables to
become Stage 2 translations, mapping from a PA to an IPA.
The amount of virtualized memory that is available is platform
dependent. For example, the zynqmp has a 40-bit range (1TB).

The range of memory can limit the number of levels that the translation
tables can walk to 3 instead of 4. This is dependent on the target platform.

If starting at Level 1, PGDs are replaced with PUD in the Vspace and Fastpath.

The virtual traslation registers are also affected by the available physical
address range because it translates from PA to IPA. The control register gets
updated accordingly.

Change-Id: Ie5033a268bc9aba34cf3287404f484b79aa2fe82