i.MX8QM 64-bit Support
This adds support for the 64-bit i.MX8 Quad Max SOC. This port does
not have support for the messaging unit, which allows the cores to communicate
with the System Control Unit (SCU). Interaction with the SCU can be
done from user space or from the bootloader before branching to the
kernel. This means that clocks and peripherals must be powered up by
the bootloader or from userspace.
32-bit is not included here either because when booting with u-boot,
the Arm Trusted Firmware (ATF) needs to be 64-bit and is compiled into
u-boot so the developer may have to alter u-boot to load a 32-bit
image.

Change-Id: I205548c53a5fe9d6990a9c20797df9c5e93ddb00