Handle SGI Group 1 Interrupts as a VCPU Fault.
The GICv3 uses system registers to inject interrupts. These
happen randomly and infrequently, and due to HCR_EL2 settings
get trapped to EL2. The trap handlers don't have a catch for
the exception, and it gets treated as a faulty instruction.

This commit adds a fault handler which checks for the
exception syndrome for trapped MRS/MSR instructions, treats
it as a VCPU fault, performs the SGI Group 1 interrupt
injection, then increments the program counter, and returns
to Linux as a handled exception.

The current setup actually injects the IRQ, which may not
be the best behavior when multiple VMs are running.
This event could also be passed into the vgic for proper handling.

Change-Id: I6a7ee8c87601f0bcd59876c5532f2df28c7e953c